refactor: 통계 계산 책임을 LottoResultService에서 Lottos로 위임
- WinningLotto에서 매칭 결과 반환
- Lottos에서 Rank 계산 담당
- 이에 맞는 테스트 코드 수정
