+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -104,4 +104,68 @@ public class WinningLottoTest {
 
         assertThat(result).isTrue();
     }
+
+    @DisplayName("하나의 로또에 대해서 로또 번호와 당첨 번호를 비교하여 순위 반환")
+    @MethodSource("provideLottos")
+    @ParameterizedTest
+    void 로또_번호와_당첨_번호를_비교하여_순위_반환(
+            List<Integer> myNumber,
+            int bonusNumber,
+            Rank expectedRank
+    ) {
+        //given
+        Lotto myLotto = new Lotto(myNumber);
+        WinningLotto winningLotto = new WinningLotto(defaultWinningNumber, new BonusNumber(bonusNumber, defaultWinningNumber));
+
+        //when
+        Rank result = winningLotto.determineRank(myLotto);
+
+        //then
+        assertThat(result).isEqualTo(expectedRank);
+    }
+
+    private static Stream<Arguments> provideLottos() {
+        return Stream.of(
+                Arguments.of(
+                        List.of(1, 2, 3, 4, 5, 6),
+                        7,
+                        Rank.FIRST
+                ),
+                Arguments.of(
+                        List.of(1, 2, 3, 4, 5, 7),
+                        7,
+                        Rank.SECOND
+                ),
+                Arguments.of(
+                        List.of(1, 2, 3, 4, 5, 10),
+                        7,
+                        Rank.THIRD
+                ),
+                Arguments.of(
+                        List.of(1, 2, 3, 4, 10, 11),
+                        7,
+                        Rank.FOURTH
+                ),
+                Arguments.of(
+                        List.of(1, 2, 3, 10, 11, 12),
+                        7,
+                        Rank.FIFTH
+                ),
+                Arguments.of(
+                        List.of(1, 2, 7, 10, 11, 12),
+                        7,
+                        Rank.FIFTH
+                ),
+                Arguments.of(
+                        List.of(1, 2, 10, 11, 12, 13),
+                        7,
+                        Rank.NONE
+                ),
+                Arguments.of(
+                        List.of(10, 11, 12, 13, 14, 15),
+                        7,
+                        Rank.NONE
+                )
+        );
+    }
 }
